[REFACTOR] 게시글 이미지 업로드 API 패키지 변경
